The implementation of updating data for an individual service has been successfully completed. TT-110.

+++ app/src/main/java/hr/foi/techtitans/ttpay/products/activity_products/UpdateServiceActivity.kt
@@ -159,7 +159,7 @@ class UpdateServiceActivity : AppCompatActivity() {
             updatedFields["description"] = descriptionEditText.text.toString()
         }
         if (isFieldChanged(priceEditText, originalPrice)) {
-            updatedFields["price"] = priceEditText.text.toString().toDouble()
+            updatedFields["price"] = priceEditText.text.toString().toInt()
         }
         if (isFieldChanged(currencyEditText, originalCurrency)) {
             updatedFields["currency"] = currencyEditText.text.toString()
